+++ Source/SourceKittenFramework/OffsetMap.swift
@@ -48,10 +48,8 @@ extension File {
     */
     private func mapOffsets(_ dictionary: [String: SourceKitRepresentable], offsetMap: OffsetMap) -> OffsetMap {
         var offsetMap = offsetMap
-        // Only map if we're in the correct file
         if let rangeStart = SwiftDocKey.getNameOffset(dictionary),
-           let rangeLength = SwiftDocKey.getNameLength(dictionary),
-           shouldTreatAsSameFile(dictionary) {
+           let rangeLength = SwiftDocKey.getNameLength(dictionary) {
             let bodyLength = SwiftDocKey.getBodyLength(dictionary) ?? 0
             let rangeMax = Int(rangeStart + rangeLength + bodyLength)
             let rangeStart = Int(rangeStart)